move constraints into interpreter
